code/requirements.txt
flask==2.1.0
google-cloud-logging==3.0.0